+++ rasa_core/training/interactive.py
@@ -256,6 +256,12 @@ def format_bot_output(
         for idx, element in enumerate(data.get("elements")):
             element_str = element_to_string(element, idx)
             output += "\n" + element_str
+
+    if data.get("quick_replies"):
+        output += "\nQuick replies:"
+        for idx, element in enumerate(data.get("quick_replies")):
+            element_str = element_to_string(element, idx)
+            output += "\n" + element_str
     return output